restructure project by moving ai client out of its own folder
