--- ai_conversation_client/src/__init__.py
@@ -1,8 +1,0 @@
-"""AI Conversation Client package.
-
-This module exposes the main AIConversationClient class.
-"""
-
-from .client import AIConversationClient
-
-__all__ = ["AIConversationClient"]
