+++ src/__init__.py
@@ -1,3 +1,4 @@
+from .ai_client import AIConversationClient
 from .issue_tracker_interface import (
     Comment,
     Issue,
@@ -7,6 +8,7 @@ from .issue_tracker_interface import (
 
 # Re-export the interfaces so users can still import from src.api
 __all__ = [
+    "AIConversationClient",
     "Comment",
     "Issue",
     "IssueTrackerClient",
