+++ src/ai_cli.py
@@ -8,7 +8,7 @@ including starting a chat, showing message history, and managing sessions.
 import argparse
 import asyncio
 
-from ai_conversation_client.src.client import AIConversationClient
+from src.ai_client import AIConversationClient
 
 
 async def interactive_chat(client: AIConversationClient, user_id: str) -> None:
