+++ tests/ai_issue_integration_test.py
@@ -1,8 +1,8 @@
 import pytest
 
-from ai_conversation_client.src.client import AIConversationClient
-from src.issue_tracker import MemoryIssueTrackerClient
+from src.ai_client import AIConversationClient
 from src.ai_issue_integration import AIIssueIntegration
+from src.issue_tracker import MemoryIssueTrackerClient
 from tests.dummy_api_client import DummyAPIClient
